Moved narrateHoveredItemInInventory to separate class
